feat(proxy): extract proxy chain UI to separate component
