+++ src/components/proxy/proxy-groups.tsx
@@ -1,22 +1,10 @@
-import { ExpandMoreRounded } from '@mui/icons-material'
-import {
-  Alert,
-  Box,
-  Chip,
-  IconButton,
-  Menu,
-  MenuItem,
-  Snackbar,
-  Typography,
-} from '@mui/material'
-import { useTheme } from '@mui/material/styles'
+import { Box } from '@mui/material'
 import { useQuery } from '@tanstack/react-query'
 import { defaultRangeExtractor, useVirtualizer } from '@tanstack/react-virtual'
 import { useLockFn } from 'ahooks'
 import {
-  type Key,
-  type MouseEvent,
-  type RefObject,
+  lazy,
+  Suspense,
   useCallback,
   useEffect,
   useLayoutEffect,
@@ -24,32 +12,36 @@ import {
   useRef,
   useState,
 } from 'react'
-import { useTranslation } from 'react-i18next'
 import { useLocation } from 'react-router'
 import { delayGroup, healthcheckProxyProvider } from 'tauri-plugin-mihomo-api'
 
 import {
   BaseEmpty,
+  BaseLoading,
   StickyVirtualList,
   StickyVirtualListHandle,
 } from '@/components/base'
 import { useProxySelection } from '@/hooks/use-proxy-selection'
 import { useVerge } from '@/hooks/use-verge'
 import { useProxiesData } from '@/providers/app-data-context'
-import { calcuProxies, updateProxyChainConfigInRuntime } from '@/services/cmds'
+import { calcuProxies } from '@/services/cmds'
 import delayManager from '@/services/delay'
 import { debugLog } from '@/utils/debug'
 
 import { ScrollTopButton } from '../layout/scroll-top-button'
 
-import { ProxyChain } from './proxy-chain'
 import {
   DEFAULT_HOVER_DELAY,
   ProxyGroupNavigator,
 } from './proxy-group-navigator'
 import { ProxyRender } from './proxy-render'
-import type { HeadState } from './use-head-state'
 import { type IRenderItem, useRenderList } from './use-render-list'
+
+const ProxyGroupsChain = lazy(() =>
+  import('./proxy-groups-chain').then((m) => ({
+    default: m.ProxyGroupsChain,
+  })),
+)
 
 function useStableCallback<T extends (...args: any[]) => any>(fn: T): T {
   const ref = useRef(fn)
@@ -63,15 +55,7 @@ interface Props {
   chainConfigData?: string | null
 }
 
-interface ProxyChainItem {
-  id: string
-  name: string
-  type?: string
-  delay?: number
-}
-
 export const ProxyGroups = (props: Props) => {
-  const { t } = useTranslation()
   const { pathname } = useLocation()
   const { mode, isChainMode = false, chainConfigData } = props
 
@@ -86,31 +70,7 @@ export const ProxyGroups = (props: Props) => {
     refetchOnReconnect: false,
   })
 
-  const [proxyChain, setProxyChain] = useState<ProxyChainItem[]>(() => {
-    try {
-      const saved = localStorage.getItem('proxy-chain-items')
-      if (saved) {
-        return JSON.parse(saved)
-      }
-    } catch {
-      // ignore
-    }
-    return []
-  })
   const [selectedGroup, setSelectedGroup] = useState<string | null>(null)
-
-  useEffect(() => {
-    if (proxyChain.length > 0) {
-      localStorage.setItem('proxy-chain-items', JSON.stringify(proxyChain))
-    } else {
-      localStorage.removeItem('proxy-chain-items')
-    }
-  }, [proxyChain])
-  const [ruleMenuAnchor, setRuleMenuAnchor] = useState<null | HTMLElement>(null)
-  const [duplicateWarning, setDuplicateWarning] = useState<{
-    open: boolean
-    message: string
-  }>({ open: false, message: '' })
 
   const { verge } = useVerge()
   const { proxies: proxiesData } = useProxiesData()
@@ -313,80 +273,18 @@ export const ProxyGroups = (props: Props) => {
     saveScrollPosition(0)
   }, [saveScrollPosition])
 
-  // 关闭重复节点警告
-  const handleCloseDuplicateWarning = useCallback(() => {
-    setDuplicateWarning({ open: false, message: '' })
-  }, [])
-
-  const currentGroup = useMemo(() => {
-    if (!activeSelectedGroup) return null
-    return (
-      availableGroups.find(
-        (group: any) => group.name === activeSelectedGroup,
-      ) ?? null
-    )
-  }, [activeSelectedGroup, availableGroups])
-
-  // 处理代理组选择菜单
-  const handleGroupMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
-    setRuleMenuAnchor(event.currentTarget)
-  }
-
-  const handleGroupMenuClose = () => {
-    setRuleMenuAnchor(null)
-  }
-
+  // 处理代理组选择
   const handleGroupSelect = (groupName: string) => {
     setSelectedGroup(groupName)
-    handleGroupMenuClose()
-
-    if (isChainMode && mode === 'rule') {
-      updateProxyChainConfigInRuntime(null)
-      localStorage.removeItem('proxy-chain-group')
-      localStorage.removeItem('proxy-chain-exit-node')
-      localStorage.removeItem('proxy-chain-items')
-      setProxyChain([])
-    }
   }
 
   const handleChangeProxy = useCallback(
     (group: IProxyGroupItem, proxy: IProxyItem) => {
-      if (isChainMode) {
-        // 使用函数式更新来避免状态延迟问题
-        setProxyChain((prev) => {
-          // 检查是否已经存在相同名称的代理，防止重复添加
-          if (prev.some((item) => item.name === proxy.name)) {
-            const warningMessage = t('proxies.page.chain.duplicateNode')
-            setDuplicateWarning({
-              open: true,
-              message: warningMessage,
-            })
-            return prev // 返回原来的状态，不做任何更改
-          }
-
-          // 安全获取延迟数据，如果没有延迟数据则设为 undefined
-          const delay =
-            proxy.history && proxy.history.length > 0
-              ? proxy.history[proxy.history.length - 1].delay
-              : undefined
-
-          const chainItem: ProxyChainItem = {
-            id: `${proxy.name}_${Date.now()}`,
-            name: proxy.name,
-            type: proxy.type,
-            delay: delay,
-          }
-
-          return [...prev, chainItem]
-        })
-        return
-      }
-
       if (!['Selector', 'URLTest', 'Fallback'].includes(group.type)) return
 
       handleProxyGroupChange(group, proxy)
     },
-    [handleProxyGroupChange, isChainMode, t],
+    [handleProxyGroupChange],
   )
 
   // 测全部延迟
@@ -520,7 +418,6 @@ export const ProxyGroups = (props: Props) => {
         onHeadState={onHeadState}
         onChangeProxy={handleChangeProxy}
         onGroupToggle={handleGroupToggle}
-        isChainMode={isChainMode}
       />
     ),
     [
@@ -528,7 +425,6 @@ export const ProxyGroups = (props: Props) => {
       handleCheckAll,
       onHeadState,
       handleLocation,
-      isChainMode,
       handleGroupToggle,
     ],
   )
@@ -547,88 +443,32 @@ export const ProxyGroups = (props: Props) => {
     [handleChangeProxy, handleCheckAll, onHeadState, handleLocation],
   )
 
-  const renderProxyList = (height: string) => (
-    <ProxyVirtualList
-      parentRef={parentRef}
-      height={height}
-      totalSize={virtualizer.getTotalSize()}
-      virtualItems={virtualItems}
-      renderList={renderList}
-      activeStickyIndex={activeStickyIndex}
-      indent={mode === 'rule' || mode === 'script'}
-      isChainMode={isChainMode}
-      measureElement={virtualizer.measureElement}
-      onLocation={handleLocation}
-      onCheckAll={handleCheckAll}
-      onHeadState={onHeadState}
-      onChangeProxy={handleChangeProxy}
-    />
-  )
-
   if (mode === 'direct') {
     return <BaseEmpty textKey="proxies.page.messages.directMode" />
   }
 
   if (isChainMode) {
-    // 获取所有代理组
-    const proxyGroups = proxiesData?.groups || []
-    const showRuleHeader = mode === 'rule' && proxyGroups.length > 0
-
     return (
-      <>
-        <Box sx={{ display: 'flex', height: '100%', gap: 2 }}>
-          <Box sx={{ flex: 1, position: 'relative' }}>
-            {showRuleHeader && (
-              <ChainRuleHeader
-                title={t('proxies.page.rules.title')}
-                selectLabel={t('proxies.page.rules.select')}
-                currentGroup={currentGroup}
-                canSelectGroup={availableGroups.length > 0}
-                onMenuOpen={handleGroupMenuOpen}
-              />
-            )}
-
-            {renderProxyList(
-              showRuleHeader ? 'calc(100% - 80px)' : 'calc(100% - 14px)',
-            )}
-            <ScrollTopButton show={showScrollTop} onClick={scrollToTop} />
-          </Box>
-
-          <Box sx={{ width: '400px', minWidth: '300px' }}>
-            <ProxyChain
-              proxyChain={proxyChain}
-              onUpdateChain={setProxyChain}
-              chainConfigData={chainConfigData}
-              mode={mode}
-              selectedGroup={activeSelectedGroup}
-            />
-          </Box>
-        </Box>
-
-        <Snackbar
-          open={duplicateWarning.open}
-          autoHideDuration={3000}
-          onClose={handleCloseDuplicateWarning}
-          anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-        >
-          <Alert
-            onClose={handleCloseDuplicateWarning}
-            severity="warning"
-            variant="filled"
-          >
-            {duplicateWarning.message}
-          </Alert>
-        </Snackbar>
-
-        <GroupSelectMenu
-          anchorEl={ruleMenuAnchor}
-          groups={availableGroups}
-          selectedGroup={activeSelectedGroup}
-          emptyText="暂无可用代理组"
-          onClose={handleGroupMenuClose}
-          onSelect={handleGroupSelect}
+      <Suspense fallback={<BaseLoading />}>
+        <ProxyGroupsChain
+          mode={mode}
+          chainConfigData={chainConfigData}
+          availableGroups={availableGroups}
+          activeSelectedGroup={activeSelectedGroup}
+          showScrollTop={showScrollTop}
+          parentRef={parentRef}
+          totalSize={virtualizer.getTotalSize()}
+          virtualItems={virtualItems}
+          renderList={renderList}
+          activeStickyIndex={activeStickyIndex}
+          measureElement={virtualizer.measureElement}
+          onCheckAll={handleCheckAll}
+          onHeadState={onHeadState}
+          onLocation={handleLocation}
+          onGroupSelect={handleGroupSelect}
+          onScrollToTop={scrollToTop}
         />
-      </>
+      </Suspense>
     )
   }
 
@@ -659,240 +499,6 @@ export const ProxyGroups = (props: Props) => {
       )}
 
       <ScrollTopButton show={showScrollTop} onClick={scrollToTop} />
-    </div>
-  )
-}
-
-type VirtualListItem = {
-  key: Key
-  index: number
-  start: number
-  end: number
-}
-
-interface ProxyVirtualListProps {
-  parentRef: RefObject<HTMLDivElement | null>
-  height: string
-  totalSize: number
-  virtualItems: VirtualListItem[]
-  renderList: IRenderItem[]
-  activeStickyIndex: number | null
-  indent: boolean
-  isChainMode?: boolean
-  measureElement: (node: Element | null) => void
-  onLocation: (group: IRenderItem['group']) => void
-  onCheckAll: (groupName: string) => void
-  onHeadState: (groupName: string, patch: Partial<HeadState>) => void
-  onChangeProxy: (
-    group: IRenderItem['group'],
-    proxy: IRenderItem['proxy'] & { name: string },
-  ) => void
-}
-
-interface ProxyGroupOption {
-  name: string
-  type: string
-  all?: unknown[]
-}
-
-interface ChainRuleHeaderProps {
-  title: string
-  selectLabel: string
-  currentGroup: ProxyGroupOption | null
-  canSelectGroup: boolean
-  onMenuOpen: (event: MouseEvent<HTMLElement>) => void
-}
-
-function ChainRuleHeader({
-  title,
-  selectLabel,
-  currentGroup,
-  canSelectGroup,
-  onMenuOpen,
-}: ChainRuleHeaderProps) {
-  return (
-    <Box sx={{ borderBottom: '1px solid', borderColor: 'divider' }}>
-      <Box
-        sx={{
-          px: 2,
-          py: 1.5,
-          borderBottom: '1px solid',
-          borderColor: 'divider',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'space-between',
-        }}
-      >
-        <Box sx={{ display: 'flex', alignItems: 'center', gap: 2 }}>
-          <Typography variant="h6" sx={{ fontWeight: 600, fontSize: '16px' }}>
-            {title}
-          </Typography>
-
-          {currentGroup && (
-            <Chip
-              size="small"
-              label={`${currentGroup.name} (${currentGroup.type})`}
-              variant="outlined"
-              sx={{
-                fontSize: '12px',
-                maxWidth: '200px',
-                '& .MuiChip-label': {
-                  overflow: 'hidden',
-                  textOverflow: 'ellipsis',
-                  whiteSpace: 'nowrap',
-                },
-              }}
-            />
-          )}
-        </Box>
-
-        {canSelectGroup && (
-          <IconButton
-            size="small"
-            onClick={onMenuOpen}
-            sx={{
-              border: '1px solid',
-              borderColor: 'divider',
-              borderRadius: '4px',
-              padding: '4px 8px',
-            }}
-          >
-            <Typography variant="body2" sx={{ mr: 0.5, fontSize: '12px' }}>
-              {selectLabel}
-            </Typography>
-            <ExpandMoreRounded fontSize="small" />
-          </IconButton>
-        )}
-      </Box>
-    </Box>
-  )
-}
-
-interface GroupSelectMenuProps {
-  anchorEl: HTMLElement | null
-  groups: ProxyGroupOption[]
-  selectedGroup: string | null
-  emptyText: string
-  onClose: () => void
-  onSelect: (groupName: string) => void
-}
-
-function GroupSelectMenu({
-  anchorEl,
-  groups,
-  selectedGroup,
-  emptyText,
-  onClose,
-  onSelect,
-}: GroupSelectMenuProps) {
-  return (
-    <Menu
-      anchorEl={anchorEl}
-      open={Boolean(anchorEl)}
-      onClose={onClose}
-      slotProps={{
-        paper: {
-          sx: {
-            maxHeight: 300,
-            minWidth: 200,
-          },
-        },
-      }}
-    >
-      {groups.map((group) => (
-        <MenuItem
-          key={group.name}
-          onClick={() => onSelect(group.name)}
-          selected={selectedGroup === group.name}
-          sx={{ fontSize: '14px', py: 1 }}
-        >
-          <Box
-            sx={{
-              display: 'flex',
-              flexDirection: 'column',
-              alignItems: 'flex-start',
-            }}
-          >
-            <Typography variant="body2" sx={{ fontWeight: 500 }}>
-              {group.name}
-            </Typography>
-            <Typography variant="caption" color="text.secondary">
-              {group.type} · {group.all?.length ?? 0} 节点
-            </Typography>
-          </Box>
-        </MenuItem>
-      ))}
-
-      {groups.length === 0 && (
-        <MenuItem disabled>
-          <Typography variant="body2" color="text.secondary">
-            {emptyText}
-          </Typography>
-        </MenuItem>
-      )}
-    </Menu>
-  )
-}
-
-function ProxyVirtualList({
-  parentRef,
-  height,
-  totalSize,
-  virtualItems,
-  renderList,
-  activeStickyIndex,
-  isChainMode,
-  measureElement,
-  onLocation,
-  onCheckAll,
-  onHeadState,
-  onChangeProxy,
-}: ProxyVirtualListProps) {
-  const theme = useTheme()
-  const stickyBackground =
-    theme.palette.mode === 'dark' ? '#1e1f27' : 'var(--background-color)'
-
-  return (
-    <div ref={parentRef} style={{ height, overflow: 'auto' }}>
-      <div style={{ height: totalSize, position: 'relative' }}>
-        {virtualItems.map((virtualItem) => (
-          <div
-            key={virtualItem.key}
-            data-index={virtualItem.index}
-            ref={measureElement}
-            style={{
-              position:
-                virtualItem.index === activeStickyIndex ? 'sticky' : 'absolute',
-              top: 0,
-              left: 0,
-              zIndex: virtualItem.index === activeStickyIndex ? 5 : undefined,
-              display:
-                virtualItem.index === activeStickyIndex
-                  ? 'flow-root'
-                  : undefined,
-              backgroundColor:
-                virtualItem.index === activeStickyIndex
-                  ? stickyBackground
-                  : undefined,
-              width: '100%',
-              transform:
-                virtualItem.index === activeStickyIndex
-                  ? undefined
-                  : `translateY(${virtualItem.start}px)`,
-            }}
-          >
-            <ProxyRender
-              item={renderList[virtualItem.index]}
-              onLocation={onLocation}
-              onCheckAll={onCheckAll}
-              onHeadState={onHeadState}
-              onChangeProxy={onChangeProxy}
-              isChainMode={isChainMode}
-            />
-          </div>
-        ))}
-        <div style={{ height: 8 }} />
-      </div>
     </div>
   )
 }
